+++ firestore-incremental-capture/functions/src/index.ts
@@ -14,7 +14,6 @@
  * limitations under the License.
  */
 
-import * as admin from 'firebase-admin';
 import * as functions from 'firebase-functions';
 
 import config from './config';
@@ -27,8 +26,6 @@ import {onBackupRestoreHandler} from './tasks/on_backup_restore_handler';
 import {runInitialSetupHandler} from './tasks/on_run_initial_setup_handler';
 import {onHttpRunRestorationHandler} from './tasks/on_http_run_restoration_handler';
 import {onFirestoreBackupInitHandler} from './tasks/on_firestore_backup_init_handler';
-
-admin.initializeApp();
 
 /**
  * Sync data to BigQuery, triggered by any change to a Firestore document