+++ firestore-incremental-capture/install/functions/build_dataflow_template.sh
@@ -1,6 +1,6 @@
 # Build Dataflow Flex Template (Download the JAR file first)
 echo -e "${YELLOW}Step 6:Building Dataflow Flex Template...${NC}"
-gcloud dataflow flex-template build gs://$PROJECT_ID.appspot.com/$EXT_INSTANCE_ID-dataflow-restore \
+gcloud dataflow flex-template build gs://$PROJECT_ID$DEFAULT_BUCKET_SUFFIX/$EXT_INSTANCE_ID-dataflow-restore \
   --image-gcr-path $LOCATION-docker.pkg.dev/$PROJECT_ID/$EXT_INSTANCE_ID/dataflow/restore:latest \
   --sdk-language JAVA \
   --flex-template-base-image JAVA11 \